+++ packages/jet-nx/src/generators/versions.ts
@@ -1,3 +1,3 @@
-export const jet = '^0.0.4';
+export const jet = '^0.0.7';
 export const afterburner = '^0.1.1';
 export const typesAwsLambda = '^8.10.83';